requirements.txt
playwright>=1.44.0
python-slugify>=8.0.1
pandas>=2.0.0
tqdm>=4.65.0
aiofiles>=23.0.0
googlesearch-python>=1.2.3
pdfplumber>=0.9.0
unidecode>=1.3.7
requests>=2.31.0
lxml>=4.9.3 
langchain==0.1.17 
langgraph==0.0.40
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest-asyncio>=0.21.0 
tiktoken>=0.5.0 
pydantic>=2.0.0 
filelock>=3.12.0 
psutil>=5.9; platform_system!="Windows"
fastapi>=0.104.0
uvicorn>=0.24.0 